+++ resources/js/components/Flickity.vue
@@ -1,6 +1,15 @@
 <template>
     <div>
-        <slot />
+        <div ref="carousel">
+            <slot></slot>
+        </div>
+        <slot
+            name="custom-ui"
+            :next="next"
+            :previous="previous"
+            :selectedIndex="selectedIndex"
+            :slides="slides"
+        ></slot>
     </div>
 </template>
 
@@ -9,10 +18,49 @@ import Flickity from "flickity"
 
 export default {
     props: {
-        options: Object
+        options: Object,
+
+        // Call resize() on Flickity instance when this prop changes to true for the first time.
+        // Useful when Flickity is initialized hidden
+        resizeOnce: Boolean,
+    },
+    data() {
+        return {
+            hasBeenResizedOnce: this.resizeOnce,
+            slides: [],
+            selectedIndex: null,
+        }
     },
     mounted() {
-        new Flickity(this.$el, this.options)
+        const vm = this
+        this.flickity = new Flickity(this.$refs.carousel, {
+            on: {
+                ready() {
+                    vm.slides = this.slides
+                    vm.selectedIndex = this.selectedIndex
+                },
+                change(index) {
+                    vm.selectedIndex = index
+                }
+            },
+            ...this.options
+        })
+    },
+    methods: {
+        next() {
+            this.flickity.next()
+        },
+        previous() {
+            this.flickity.previous()
+        },
+    },
+    watch: {
+        resizeOnce(shouldResize) {
+            if (!shouldResize) return
+            if (this.hasBeenResizedOnce) return
+
+            this.flickity.resize()
+        },
     },
 }
 </script>