+++ tailwind.config.js
@@ -20,11 +20,19 @@ module.exports = {
     extend: {
       colors: {
         gray: {
-          200: '#EDEDED',
+          200: '#ededed',
+          300: '#cdcdcd',
           500: '#969696',
           600: '#777',
+          800: '#333',
         },
       },
+      spacing: {
+        '104': '26rem',
+      },
+      textDecorationThickness: {
+        3: '3px',
+      }
     },
   },
   plugins: [],