+++ resources/js/components/TabPanel.vue
@@ -1,16 +1,15 @@
 <template>
-    <div v-cloak v-show="controller.activeIndex === index">
-        <slot></slot>
+    <div v-show="controller.activeIndex === index">
+        <slot :active="controller.activeIndex === index"></slot>
     </div>
 </template>
 
 <script>
-
 export default {
     inject: {
-      controller: {
-        from: 'tabsController'
-      }
+        controller: {
+            from: "tabsController",
+        },
     },
     data() {
         return {