+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - use authority matcher in item indexing
 
+### Changed
+- delete formerly attributed authors
+- do not harvest formerly attributed authors
+
+
 ## [2.43.0] - 2022-02-04
 ### Added
 - basic Tailwind set-up