+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file[^1].
 - "frameColor" to JIRA request in order process
 - use authority matcher in item indexing
 - "latest content" section for new homepage
+- import multiple images from pnp karasek
 
 ### Changed
 - delete formerly attributed authors