+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.44.0] - 2022-02-01
+### Added
+- "frameColor" to JIRA request in order process
+- use authority matcher in item indexing
+- "latest content" section for new homepage
+- import multiple images from pnp karasek
+
+### Changed
+- delete formerly attributed authors
+- do not harvest formerly attributed authors
+
 ## [2.43.0] - 2022-02-04
 ### Added
 - basic Tailwind set-up